Backend Server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
pymongo==4.6.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2
httpx==0.25.1
qrcode==7.4.2
python-jose==3.3.0 